+++ 4-period/java-vendingmachine-precourse-main/src/main/java/vendingmachine/domain/price/InputPrice.java
@@ -5,12 +5,12 @@ import static vendingmachine.exception.ErrorMessage.INVALID_INPUT_AMOUNT;
 import vendingmachine.exception.CustomIllegalArgumentException;
 
 public class InputPrice extends Price {
-    public InputPrice(final long amount) {
+    public InputPrice(final int amount) {
         super(amount);
         validate(amount);
     }
 
-    private void validate(final long amount) {
+    private void validate(final int amount) {
         if (amount == 0) {
             throw new CustomIllegalArgumentException(INVALID_INPUT_AMOUNT);
         }
