+++ 4-period/java-vendingmachine-precourse-main/src/main/java/vendingmachine/domain/price/coin/CoinGenerator.java
@@ -6,5 +6,5 @@ import vendingmachine.domain.price.Price;
 @FunctionalInterface
 public interface CoinGenerator {
 
-    Map<Coin, Long> generateCoins(Price price);
+    Map<Coin, Integer> generateCoins(Price price);
 }
